+++ app/build.gradle
@@ -9,11 +9,11 @@ plugins {
 }
 
 android {
-	namespace 'com.example.open_realms'
+	namespace 'at.vunfer.openrealms'
 	compileSdk 33
 
 	defaultConfig {
-		applicationId "com.example.open_realms"
+		applicationId "at.vunfer.openrealms"
 		minSdk 29
 		targetSdk 33
 		versionCode 1
@@ -80,8 +80,9 @@ dependencies {
 	implementation 'com.google.android.material:material:1.8.0'
 	implementation 'androidx.constraintlayout:constraintlayout:2.1.4'
 
-	testImplementation 'org.junit.jupiter:junit-jupiter-api:5.7.0'
-	testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine:5.7.0'
+	testImplementation 'org.junit.jupiter:junit-jupiter-api:5.9.2'
+	testImplementation 'org.testng:testng:6.9.6'
+	testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine:5.9.2'
 
 	androidTestImplementation 'androidx.test.ext:junit:1.1.5'
 	androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.1'
