+++ app/src/main/java/at/vunfer/openrealms/presenter/PlayAreaPresenter.java
@@ -1,0 +1,4 @@
+/* Licensed under GNU GPL v3.0 (C) 2023 */
+package at.vunfer.openrealms.presenter;
+
+public class PlayAreaPresenter {}
