--- app/src/main/java/com/example/open_realms/presenter/MarketPresenter.java
@@ -1,4 +1,0 @@
-/* Licensed under GNU GPL v3.0 (C) 2023 */
-package com.example.open_realms.presenter;
-
-public class MarketPresenter {}
