+++ app/src/test/java/at/vunfer/openrealms/ExampleUnitTest.java
@@ -1,0 +1,18 @@
+/* Licensed under GNU GPL v3.0 (C) 2023 */
+package at.vunfer.openrealms;
+
+import static org.junit.Assert.*;
+
+import org.testng.annotations.Test;
+
+/**
+ * Example local unit test, which will execute on the development machine (host).
+ *
+ * @see <a href="http://d.android.com/tools/testing">Testing documentation</a>
+ */
+public class ExampleUnitTest {
+    @Test
+    public void addition_isCorrect() {
+        assertEquals(4, 2 + 2);
+    }
+}
